tasks/cephfs: add function to configure guest credentials
In the test_volume_client module, refactor code to add a function
that configures credentials for a guest to mount a volume. The
function could be used by tests to setup a guest client to mount
a volume with fewer lines of code.

Signed-off-by: Ramana Raja <rraja@redhat.com>
(cherry picked from commit 02fc6d6acc1246881c173d693773658b210f1ad3)