osd/PG: set last_* PG fields on a creating PG
Use the value from pg_history_t that ultimately came from the
mon if last_epoch_started == 0.  This establishes a sane lower
bound on these timestamps, and prevents a mon health error from
coming up about how a PG is stuck in whatever state while it is
being created.

(We addressed half of this problem in commit 6ca6aeac, but the
warning comes back as soon as the OSD reports with a creating
state.)

Fixes: http://tracker.ceph.com/issues/14952
Signed-off-by: Sage Weil <sage@redhat.com>