osd/PG: exclude removed osd from down set of PriorSet
The PriorSet constructor will allow a nonexistent OSD
go into the down set, which is not quite appropriate.

E.g., on receiving a new osdmap, we'll call affected_by_map()
again to check if the given map affects the PriorSet and
will restart Peering process if it is. As affected_by_map()
relies on the down set and the blocked_by set to react
properly, including a perviously nonexistent OSD in the down
set can be rather misleading sometimes, although it may be
rare.

Signed-off-by: xie xingguo <xie.xingguo@zte.com.cn>