tools/rbd: introduce a new field of visible in Action Class
The new field means whether we want to show this Action in
command help. Sometimes, we want to make one action to be
deprecated and invisible to user, but we have to keep it
working for compatibility.

Then we need to support this action but make it invisible.

Signed-off-by: Dongsheng Yang <dongsheng.yang@easystack.cn>
(cherry picked from commit 367a5fe507fd80750727e9efd92c90e67967ae35)