tools/rbd: introduce a simple bench for read
Currently, we have a bench-write for in rbd tool, but we don't have
a simple bench for read.

This patch add a new subcommand of bench, with an option of --io-type.
Then we can do a simple bench for write or read.

Signed-off-by: Dongsheng Yang <dongsheng.yang@easystack.cn>
(cherry picked from commit 23a135d2a8a9fc5acb8d833419d564eac8f3933b)