osd: make health check more accurate for heartbeat mechanism
The original logic will reuse the timestamp which we send a PING to
the specific heartbeat peer to update the last_rx_front[back] field
on receiving the corresponding REPLIES, which later shall be honoured
as the exact timestamps we succeed in getting the corresponding
REPLIES and used to calculate the heartbeat latency and determine
whether the relevant peer is dead.

However this is not accurate enough as there may be a delay between
we receive a PINGREPLY and call heartbeat_check(), which is
periodically scheduled by the tick() thread and heartbeat_thread,
to calculate the real heartbeat latency:

1. the tick() thread will wake up and call OSD::heartbeat_check()
   at 1s interval, but can be blocked and delayed by the big osd_lock.
2. the heartbeat_thread will wake up and call OSD::heartbeat_check()
   at a random-choosen interval, which can vary from 0.5s ~6.5S, and
   is blockless, since it requires the heartbeat_lock only.
3. the heartbeat_thread will also be responsible for sending PINGs
   each time it wakes up.

With the prior knowledge above, the following example might be a
little more comprehensible now:

1. we send ping at 2016/5/13 14:59:10
2. we get both replies at 2016/5/13 14:59:11 and mark last_rx_* as
   2016/5/13 14:59:10(NB: we always update last_rx_* fields to the
   timestamp we sending pings, instead of receiving them)
3. heartbeat_thread wakes up at 2016/5/13 14:59:15 and calls
   OSD::heartbeat_check() for the heartbeat latency.
   However, it will figure out that the latency is

     now(2016/5/13 14:59:15) - last_rx_*(2016/5/13 14:59:10) = 5s

   instead of

     2016/5/13 14:59:11 - 2016/5/13 14:59:10 =  1s

   which is obviously wrong.
4. further, heartbeat_thread resends a new ping at 2016/5/13 14:59:15
   meanwhile and the peer is now unable to respond such a call.
5. the tick() thread keeps ticking and the heartbeat latency now may
   appear to be the following series:
     now(2016/5/13 14:59:16) - last_rx_*(2016/5/13 14:59:10) = 6s
     now(2016/5/13 14:59:17) - last_rx_*(2016/5/13 14:59:10) = 7s
     now(2016/5/13 14:59:18) - last_rx_*(2016/5/13 14:59:10) = 8s
     ...
   this is also **not** accurate as we know we send new ping at
   2016/5/13 14:59:15 and shall count down from then.
6. However the last_rx_* fields shall still be consider as the
   lower bound we can't hear a reply from the front/back side.

As both heartbeat interval(osd_heartbeat_interval, default to 6)
and grace(osd_heartbeat_grace, default to 20) are configurable
options, the above case may become fatal if we set a longer
osd_heartbeat_interval or a shorter osd_heartbeat_grace and thus
shall be resolved.

Because we keep sending new pings periodically, the old pings
could be deprecated by the newly sent pings, thus we introduces a map
to keep tracking of the ping-history here, which consists of
three elements:

1. "tx_time", worked as map key, indicates the exact timestamp we send
   a ping.
2. "deadline", indicates we shall receive all replies until then,
   otherwise we consider this peer as "dead".
3. "rx_refs_remaining", indicates how many replies for the corresponding
   ping are still in-flight. The initial value is 2(the front side and the
   back side).

We insert an item into the map above every time we send a ping, and
drop the "rx_refs_remaining" by 1 each time we get a reply for the
tracked ping. If "rx_refs_remaining" drops to 0, we know all the replies
have been successfully collected and we can safely erase the relevant
item from the map as well as older ones if there is any.

By comparing the current timestamp with the oldest deadline, we can now
make a much accurate decision about whether the corresponding peer is
healthy or not. And by setting last_rx_* to the timestamp we receiving
the reply, the lower bound that we can't hear a reply from the corresponding
connection is also much clear now.

Signed-off-by: xie xingguo <xie.xingguo@zte.com.cn>