rpm: lean_debug_symbols bcond
The line added by bd46281 causes -g1 to be passed to gcc instead of -g.
As desirable as this may be in some cases, we need to put it in a %bcond
so it can be enabled only when needed.

Signed-off-by: Nathan Cutler <ncutler@suse.com>
(cherry picked from commit 564c7513777ae4409f0e8c66d7686d07a3bab7a4)