rpm: use gcc manpage terminology for debug info bcond
In its description of the -g option, the gcc manpage speaks of "debugging
information", not "debug symbols". And for -g1 the manpage says "Level 1
produces minimal information, enough for making backtraces in parts of the
program that you don't plan to debug. This includes descriptions of functions
and external variables, and line number tables, but no information about local
variables."

The name "minimal_debugging_information" is also easier to understand.

Signed-off-by: Nathan Cutler <ncutler@suse.com>