librados: add sanity check for osd[pg]_command
We can't just return if we fail to execute these commands,
because we pass a callback in, which won't be executed on
error exit and normally it is the caller's responsibility
to release it.

So we assert instead of normal error handling here.
NB: this won't be a problem currently as osd[pg]_commands
always return 0.

Signed-off-by: xie xingguo <xie.xingguo@zte.com.cn>