mds/server: fix rare race when waitting for osdmap
If wait_for_map() returns true, we have got the requested osdmap
at the specified epoch, so we shall try again to do further
verification whether we have the specific pool or not.

The above case can happen because we drop the objecter internal
rwlock during the switch between the objecter->with_osdmap()
and objecter->wait_for_map() methods.

Signed-off-by: xie xingguo <xie.xingguo@zte.com.cn>