librbd: journal callback to interrupt replay
If the exclusive lock is lost while the journal is in the
replay state, the journal close will block until the replay
completes. The new callback will facilitate stopping replay
in a timely fashion so that the journal can be closed.

Signed-off-by: Jason Dillaman <dillaman@redhat.com>
(cherry picked from commit 9687e5e34aca98934fcf04089ead2794629455a1)