doc: remove btrfs contradiction
btrfs is not "the ideal choice for Ceph clusters" because it isn't mature and
the promised features (like deduplication) have not materialized.

The "Filesystems -> Recommended" section of this very same document (several
paragraphs) above the paragraph being modified by this commit, has been
modified to say "We used to recommend btrfs for testing, development, and any
non-critical deployments. . . . However . . . ."

This PR removes this contradiction.

Signed-off-by: Nathan Cutler <ncutler@suse.com>