cls: adding bucket count to cls_user_types
We now maintain the bucket_count in cls_user_bucket_entry and
total_buckets in cls_user_stats. Storing them in cls_user_stats allows
us to get the total bucket count from the omap header itself instead of
doing a call to list all the keys in order to do a count. This can be
used to enforce the bucket quota

Signed-off-by: Abhishek Lekshmanan <abhishek@suse.com>