osd: limit omap data in push op
We already have the config osd_recovery_max_chunk to limit the total
size of omap entries and omap data. But we need an individual config
to limit the number of omap entries independently. We call this config
osd_recovery_max_omap_entries_per_chunk here with the default number
of 64000.

Signed-off-by: Wanlong Gao <wanlong.gao@easystack.cn>