rgw: fix the handling of rgw_swift_url_prefix.
This patch fixes to the support for placing the Swift API in the root
of URL hierarchy. Unfortunately, the whole concept exhibits a severe side
effect: inability to deploy RadosGW in multi-site configuration.

The sole reason behind this fix is the fact we claimed in documentation
that the feature is available.

Fixes: http://tracker.ceph.com/issues/16673
Signed-off-by: Radoslaw Zarzynski <rzarzynski@mirantis.com>
(cherry picked from commit dbf5b5b73dc5ceb20959a4a55dd695cd5f95d034)
Signed-off-by: Matt Benjamin <mbenjamin@redhat.com>