os/bluestore: simplify and fix blob invalidation
We can't invalidate based on any particular lextent we drop, as there may
be other lextents (in this onode or in another one) that reference the
same data.  This can break even a single multi-write transaction that
relies on the buffer cache to fill out blocks.

Fix by just doing the unallocated version unconditionally.

Signed-off-by: Sage Weil <sage@redhat.com>