Fix the Ceph and Ceph UT compiling error in AArm64
When compile the Ceph, because some compiling options are
missed, the compiler stops at src/erasure-code.
Update the related CMakelists.txt files, and make the
Ceph pass the compiling.

When compile the Ceph UT, because the -msse and -msse2
are hardcoded in the test/CMakeLists.txt, the Ceph UT build
fail in AArm64. Update the source code to fix the issue.

In Ceph erasure UT, because the usful libary is missed, the
link is broken. Update test/erasure-code/CMakeLists.txt to
fix the link error.

Signed-off-by: tone.zhang <tone.zhang@linaro.org>