For civetweb: parse and accept an optional trailing 's' on port number.
To do https: append an 's' to portno.  The code that actually
passes that to civetweb is elsewhere.  But apparently we need
to set "port" in the frontend here because logic elsewhere
in radosgw needs to know on which port a connection was accepted,
and there's no logic in civetweb to indicate that.

Signed-off-by: Marcus Watts <mwatts@redhat.com>