osd/ReplicatedPG: do not whiteout non-existent obj when reading it
When doing list_snaps on an non-existent object in base pool, a whiteouted
object is created in the writeback cache tier, and an empty snapset is
attached to it. Then the list_snaps op succeeds and gets the empty snapset,
which makes librbd::DiffIterate work improperly.

This patch add return value of list_snaps to ReplicatedPG::CopyResults, and
also pass op->may_write() to ReplicatedPG::finishi_promote. A whiteouted
object is only created when the head object is not exist and if list_snaps
doesn't fail or the original op is write.

Fixes: http://tracker.ceph.com/issues/16890
Signed-off-by: Zhao Chao <zhaochao1984@gmail.com>