Build civetweb with ipv6 support.
This fixes a regression where USE_IPV6 is no longer defined.
Formerly, this was set in a terribly confusing and weird way.
The civetweb source was customized with a file "include/civetweb_conf.h"
that defined USE_IPV6.  This file isn't used or referenced anywhere in
the civetweb source (or build).  The ceph automake build has its own
separate process to make civetweb.o, which doesn't use any part
of the build logic in civetweb.  Instead, it builds civetweb.o
with its own rule passing in the compiler option
"--include $(srcdir)/civetweb/include/civetweb_conf.h".

This is both silly and confusing, plus it makes it harder to update
the civetweb source.  Instead of passing in --include X, we can just
pass in -DUSE_IPV6.  In cmake that gets set using COMPILE_DEFINITIONS.
Cmake doesn't support doing that on a per-file basis, but
it's perfectly happy doing that for an object library containing one
object.  This is not coincidentally just exactly how we're already
building civetweb.c.o with cmake.

Signed-off-by: Marcus Watts <mwatts@redhat.com>