os/bluestore: do not track bluefs_extents via freelist
We explicitly track bluefs extents in the bluefs_extents
interval_set stored in the superblock.  Telling the
freelist about it is redundant.  It's also slow, because
bluefs uses a bit portion of the device and the bitmap
representation has to flip a lot of bits in the KV store
during mkfs (or possibly during a freespace balance).

Signed-off-by: Sage Weil <sage@redhat.com>