osd/osd_type: check if pool is gone during check_new_interval()
In is_new_interval() we'll return true if the relevant pool
does not exist in the lastmap, which means theoretically
we could access violation here if we always suppose the pool
did exist in the lastmap.

This PR add a check to make sure that the above case
does not happen.

Signed-off-by: xie xingguo <xie.xingguo@zte.com.cn>