client: pass a mask parameter to path_walk
ll_walk expects to get a set of attributes out of a path_walk. Pass a
caps mask parameter into path_walk, and then apply it when we reach the
last component of the path.

This may prevent us from having to further iteract with the server after
the pathwalk, in some cases. If we know that we're going to need certain
caps to do the actual operation we can request them during the lookup
and may have all that we need by the time we go to do the real request.

Signed-off-by: Jeff Layton <jlayton@redhat.com>