systemd: Flush OSD Journal after stop
After stopping a OSD flush its journal to the data store. This makes sure
the data store holds a full copy of the OSD's data without a need of the journal
device.

When using external journals this makes the OSD disks more hot-pluggable between
different nodes.

On a different node one can simply create a blank journal for the OSD and run the
OSD again.

Signed-off-by: Wido den Hollander <wido@42on.com>