Librados: Fix deadlock in watch_flush
In previous code, in the watch_flush, it is waiting on conditon
with holding the "lock". The condition will only be signal by
finisher thread, but sadly, in some cases,when finisher queue
is not empty, some context need to take the "lock", thus deadlock.

To avoid concurrent call to shutdown cause race condition, add
a shutdown_lock to ensure only one concurrent in shutdown function.

Signed-off-by: Xiaoxi Chen <xiaoxchen@ebay.com>