test: timeout verification that mon is unreachable
Without a timeout on the command, it may hang for a very long time,
hunting for new mons. If it hangs for more than 60 seconds, it is
safe to assume the mon is indeed down.

Fixes: http://tracker.ceph.com/issues/16477

Signed-off-by: Loic Dachary <loic@dachary.org>
(cherry picked from commit a5e5119bd138a97f35737c2b50d1e621fa8286d6)