test_memstore_clone: use StoreTestFixture to do the setup/teardown
rename test case name from `MemStore` to `MemStoreClone`, because
if we use `TEST_F` to group the tests into a fixture, `MemStore`
will be used as the class name of the fixture, and hence conflicts
with the existing `MemStore` class under testing.

but the object store is created and destroyed for every test. before
this change, we do this for every test case. if it's a concern
we need to use SetUpTestCase() and TearDownTestCase() instead.

Signed-off-by: Kefu Chai <kchai@redhat.com>