librados: only request ack if AioComplete has an ack callback
Probalby no librados users are actually using the ack callback, but we are
still requesting an ack message from the OSD.  Only request it if the
AioCompletion has a callback defined for it.

This might make users who are no setting the callback but are install
polling the completion or using wait_for_complete() unhappy, but I don't
think they actually exist.  And even if they do, they are highly unlikely
to notice the difference given the way the OSDs are implemented.

Signed-off-by: Sage Weil <sage@redhat.com>