osd: add an option "osd_ignore_bad_map_crc" to ignore mismatched crc
add an option named "osd_ignore_bad_map_crc" for OSD.

the OSD will request for full maps from monitor if the CRC of the
resulting full map after applying the incremental map does not match
with the one stored in inc map being applied. in most cases, this
behaviour is just annoying but harmless. it just adds up the burden
of monitor add cause extra traffic in the network. but in a large
deployment, after upgrading the monitor with change(s) in the encoding
of OSDMap, any changes related to osd map, the monitors is obliged to
reply the full map request from old clients which are not able to
re-encode the full map with the expected CRC, this incurs unnecessary
overhead of the network and the load on both peers. so we can choose
to disable this functionality on the client side while  upgrading
the cluster.

Fixes: http://tracker.ceph.com/issues/17386
Signed-off-by: Kefu Chai <kchai@redhat.com>