osd: add an option "osd_map_request_full_on_boot"
if OSD does not request for full map at seeing mismatched CRC of
re-encoded osd map, the osd maps on the OSD and monitor will diverge
over time. this could cause mysterious problem because of the
divergence in the cluster. so we need to enable this option and reboot
the osd if `osd_ignore_bad_map_crc` is enabled when upgrading this OSD
in question.

Fixes: http://tracker.ceph.com/issues/17386
Signed-off-by: Kefu Chai <kchai@redhat.com>