src/osd: relax the requirement that we scrub a whole hash value
Previously, we needed to scrub all objects in clones in a single
hash value mainly to ensure that _scrub had access to all clones
of a single object at the same time.  Instead, just avoid letting
head or snapdir be a boundary (see the comment in the commit
for details).

Signed-off-by: Samuel Just <sjust@redhat.com>