qa/workunits/rbd: check status also in pool dir after asok commands
wait_for_image_replay_stopped returns not when the state is stopped,
but when the state is not replaying. So a race was possible when an
asok command was running when the previos stop command was still in
progress, leading to unexpected results.

Signed-off-by: Mykola Golub <mgolub@mirantis.com>
(cherry picked from commit 2882f8c01a42ee3cfe22a838b64f21c93cf84d16)

Conflicts:
	qa/workunits/rbd/rbd_mirror.sh (tests for not backported features)